auth_server/requirements.txt
redis
fastapi
authlib
uvicorn[standard]
python-dotenv